backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
apscheduler==3.10.4
requests==2.31.0
openpyxl==3.1.2
reportlab==4.2.2
jinja2==3.1.4
python-multipart==0.0.9
pydantic-settings==2.2.1
pytz==2024.1
